Accept / process orders is working but needs error testiing
I need to get transaction working but shouldn't be hard. UI accepts an
order -> warehouse createsOrder on 1st loop and then uses continueOrder
on additional products to the same order.  Process order is a private
function in warehouse.  I had to change client, invoice, invoiceentry,
order, product, and waitlistentry an addition to UI and warehouse.